Update: unify L3-L2 example entry points as test_*.py under examples/
